Final update

+++ src/utils.cpp
@@ -1,5 +1,5 @@
 #include "utils.h"
 
-QString Utils::moneyDisplay(int cents) {
+QString Utils::moneyDisplay(double cents) {
   return QString("$") + QString::number(cents / 100.0, 'f', 2);
 }
